refactored discord_app
